chore: test refactoring
